+++ R/cbs_download_table.R
@@ -34,7 +34,10 @@ cbs_download_table <- function( id
                           , typed = FALSE
                           , base_url = getOption("cbsodataR.base_url", BASE_URL)){
   meta <- cbs_download_meta(id=id, dir=dir, cache=cache, base_url = base_url)
-
+  
+  # check if filter values are valid
+  f <- get_filter(..., .meta = meta)
+  
   cbs_download_data( id   = id
                    , path = file.path(dir, "data.csv")
                    , ...