+++ R/filter.R
@@ -1,11 +1,21 @@
 # creates a string that can be used in a $filter query
-column_filter <- function(column, values){
+column_filter <- function(column, values, allowed = NULL){
   if (is.character(values)){
+    if (is.character(allowed)){
+      valid <- values %in% allowed
+      if (!all(valid)){
+        warning("Value(s): "
+               , paste0("'", values[!valid],"'", collapse = ", ")
+               , " are not a valid selection for '",column, "'. "
+               , "Consult the meta data."
+               , call. = FALSE
+               )
+      }
+    }
     query <- eq(values, column)
   } else if (is_query(values)){
       query <- values
       query$column <- column
-    
   } else {
     stop("Unsupported query: '", values, ".'")
   }
@@ -14,13 +24,16 @@ column_filter <- function(column, values){
 
 # creates a filter string that can be used in $filter query
 # usage: get_filter(Periode=c(2001))
-get_filter <- function(..., filter_list=list(...)){
+get_filter <- function(..., filter_list=list(...), select = NULL, .meta = NULL){
   if (length(filter_list) == 0){
     return(NULL)
   }
-  #browser()
   query <- sapply(names(filter_list), function(column){
-    filter <- column_filter(column, filter_list[[column]])
+    filter <- column_filter( column
+                           , filter_list[[column]]
+                           # check if values are allowed
+                           , allowed = .meta[[column]][["Key"]]
+                           )
     paste0("(", as.character(filter, column = column), ")")
   })